Refactoring header and improve routes
