--- src/MyList.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const MyList = () => (
-  <section className="section">
-    <div className="container">
-      <h1 className="title">minha lista</h1>
-    </div>
-  </section>
-)
-
-export default MyList;
