+++ src/pages/MyList/Index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const MyList = () => (
+  <div>
+    <h1 className="title">minha lista de filmes</h1>
+  </div>
+)
+
+export default MyList;
