+++ src/pages/MyList/New.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const MyListNew = () => (
+  <div>
+    <h1 className="title">adicionar um filme à lista</h1>
+  </div>
+)
+
+export default MyListNew;
